+++ fluxo-kmp-conf/src/main/kotlin/fluxo/conf/impl/ProjectUtils.kt
@@ -1,5 +1,7 @@
 package fluxo.conf.impl
 
+import com.sun.management.OperatingSystemMXBean
+import java.lang.management.ManagementFactory
 import org.gradle.api.Action
 import org.gradle.api.Project
 import org.gradle.api.artifacts.dsl.DependencyHandler
@@ -7,7 +9,26 @@ import org.gradle.api.plugins.AppliedPlugin
 import org.gradle.api.plugins.PluginAware
 
 
-internal val CPUs = Runtime.getRuntime().availableProcessors()
+internal val CPUs: Int = Runtime.getRuntime().availableProcessors()
+
+/** The maximum amount of memory that the Java virtual machine will attempt to use. */
+internal val XMX: Long = Runtime.getRuntime().maxMemory()
+
+internal val TOTAL_OS_MEMORY: Long = run {
+    try {
+        ManagementFactory.getPlatformMXBean(OperatingSystemMXBean::class.java)?.apply {
+            return@run try {
+                @Suppress("DEPRECATION")
+                totalPhysicalMemorySize
+            } catch (_: Throwable) {
+                @Suppress("Since15")
+                totalMemorySize
+            }
+        }
+    } catch (_: Throwable) {
+    }
+    return@run -1
+}
 
 
 internal val Project.isRootProject: Boolean